+++ src/main/kotlin/com/example/entities/product/ProductDetails.kt
@@ -6,7 +6,7 @@ data class ProductDetails(
     val id: Int,
     var name: String,
     var description: String,
-    var price: Long,
+    var price: Double,
     var recommendations: String,
     var url: String,
     var subcategoryId: Int,
