+++ src/main/kotlin/com/example/entities/product/ProductDraft.kt
@@ -5,7 +5,7 @@ import com.example.entities.opinion.Opinion
 data class ProductDraft(
     var name: String,
     var description: String,
-    var price: Long,
+    var price: Double,
     var recommendations: String,
 //    var opinions: List<Opinion>
 )
